requirements.txt
anyio==3.6.2
Brotli==1.0.9
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mutagen==1.46.0
pycryptodomex==3.17
python-dotenv==1.0.0
python-telegram-bot==20.2
rfc3986==1.5.0
sniffio==1.3.0
websockets==11.0.2
yt-dlp==2023.3.4
